Reorganize and clarify plugin code
I did this to help myself understand what is going 
on in the plugin. Moving the settings load up 
before the class definition is a pattern I've seen 
in other ST2 plugins. Using "constants" for the 
settings keys and expanding the variable names to 
clarify intent. Last but not least, a somewhat 
silly obsession with the line character limit.